Digital KU Payment System
This is a web development project where we worked on a part of the development of a system.The system is for making all the aspects of Khulna University digital (web based). The payment system is for online marketing from the library or from the stall and paying it through the web. The project is developed on the platform called "simtier"